grafting branches

+++ CAD/Line.pde
@@ -1,9 +1,9 @@
 class Line extends Shape {
   
-  float x2, y2, z2;
+  int x2, y2, z2;
 
-  Line(float xx1, float yy1, float zz1, 
-  float xx2, float yy2, float zz2, int mod) {
+  Line(int xx1, int yy1, int zz1, 
+  int xx2, int yy2, int zz2, int mod) {
     super(xx1, yy1, zz1, mod);
     x2 = xx2;
     y2 = yy2;
